MyBatis分页BUG

+++ gzz-retail/src/main/java/com/gzz/retail/infra/persistence/pojo/PAttributePo.java
@@ -29,5 +29,5 @@ public class PAttributePo implements Serializable  {
 	/**
 	 *  排序 idx
 	 */
-	private Byte idx;
+	private int idx;
 }
